Add tzdata and time_zone module
Game servers must declare the time zone within which they operate. The
client expects to receive this information as an offset from UTC in
hours.

Gleam doesn't seem to have a time zone package that is kept up-to-date
with the frequently-updated IANA tzdata database, so this pulls in the
tzdata Elixir package, which updates itself at runtime.